#132 refactor: atualização do código antigo
