+++ src/factories/useCases/createUser.ts
@@ -1,9 +1,9 @@
-import { BcryptAdapter } from '../../adapters/bcryptAdapter'
+import { BcryptService } from '../../infrastructure/service/bcrypt.service'
 import UserRepository from '../../repository/userRepository'
 import { CreateUserUseCase } from '../../useCase/createUser/createUserUseCase'
 
 export const makeCreateUser = () => {
-  const encryptor = new BcryptAdapter()
+  const encryptor = new BcryptService()
   const userRepository = new UserRepository()
   return new CreateUserUseCase(encryptor, userRepository)
 }
