+++ src/factories/useCases/updateUser.ts
@@ -1,9 +1,9 @@
-import { BcryptAdapter } from '../../adapters/bcryptAdapter'
+import { BcryptService } from '../../infrastructure/service/bcrypt.service'
 import UserRepository from '../../repository/userRepository'
 import { UpdateUserUseCase } from '../../useCase/updateUser/updateUserUseCase'
 
 export const makeUpdateUser = () => {
   const userRepository = new UserRepository()
-  const encryptor = new BcryptAdapter()
+  const encryptor = new BcryptService()
   return new UpdateUserUseCase(userRepository, encryptor)
 }
